refactor: change filename patterns
