+++ src/machines/controllers/machine.controller.ts
@@ -1,9 +1,9 @@
 import { Controller, Get, Post, Patch, Delete, Body, Param } from '@nestjs/common';
 
-import { Machine } from '../../machines/interfaces/machine.interface';
+import { Machine } from '../interfaces/machine.interface';
 import { Establishment } from '../../establishments/interfaces/establishment.interface';
-import { MachinesService } from '../services/machines.service';
-import { MachineDTO } from '../dto/machines.dto';
+import { MachinesService } from '../services/machine.service';
+import { MachineDTO } from '../dto/machine.dto';
 import { MachineConfig } from '../machine.config';
 
 @Controller(MachineConfig.controllerDefinition)
