+++ src/machines/dto/machine.dto.ts
